pyproject.toml
[project]
name = "gdg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "google-auth>=2.49.1",
    "google-genai>=1.68.0",
    "requests>=2.33.0",
]
